@@ -7,8 +7,8 @@ struct NavigationCoordinatableView<T: NavigationCoordinatable>: View {
     var coordinator: T
     private let id: Int
     private let router: NavigationRouter<T>
-    @ObservedObject var presentationHelper: PresentationHelper<T>
-    @ObservedObject var root: NavigationRoot
+    @StateObject var presentationHelper: PresentationHelper<T>
+    @StateObject var root: NavigationRoot
     
     var start: AnyView?
     
@@ -94,10 +94,6 @@ struct NavigationCoordinatableView<T: NavigationCoordinatable>: View {
     init(id: Int, coordinator: T) {
         self.id = id
         self.coordinator = coordinator
-        self.presentationHelper = PresentationHelper(
-            id: self.id,
-            coordinator: coordinator
-        )
         
         self.router = NavigationRouter(
             id: id,
@@ -107,8 +103,10 @@ struct NavigationCoordinatableView<T: NavigationCoordinatable>: View {
         if coordinator.stack.root == nil {
             coordinator.setupRoot()
         }
-        
-        self.root = coordinator.stack.root
+
+        let helper = PresentationHelper(id: id, coordinator: coordinator)
+        self._presentationHelper = StateObject(wrappedValue: helper)
+        self._root = StateObject(wrappedValue: coordinator.stack.root)
         
         RouterStore.shared.store(router: router)
         
